+++ src/lib/proof-system/proof.ts
@@ -1,8 +1,12 @@
-import { initializeBindings, withThreadPool } from '../../snarky.js';
+import {
+  areBindingsInitialized,
+  initializeBindings,
+  withThreadPool,
+} from '../../snarky.js';
 import { Pickles } from '../../snarky.js';
 import { Field, Bool } from '../provable/wrapped.js';
 import type {
-  FlexibleProvablePure,
+  FlexibleProvable,
   InferProvable,
 } from '../provable/types/struct.js';
 import { FeatureFlags } from './feature-flags.js';
@@ -25,8 +29,8 @@ type MaxProofs = 0 | 1 | 2;
 type ProofClass = Subclass<typeof ProofBase>;
 
 class ProofBase<Input = any, Output = any> {
-  static publicInputType: FlexibleProvablePure<any> = undefined as any;
-  static publicOutputType: FlexibleProvablePure<any> = undefined as any;
+  static publicInputType: FlexibleProvable<any> = undefined as any;
+  static publicOutputType: FlexibleProvable<any> = undefined as any;
   static tag: () => { name: string } = () => {
     throw Error(
       `You cannot use the \`Proof\` class directly. Instead, define a subclass:\n` +
@@ -55,8 +59,8 @@ class ProofBase<Input = any, Output = any> {
    */
   declare() {
     if (!ZkProgramContext.has()) return false;
-    const Proof = this.constructor as Subclass<typeof ProofBase>;
-    ZkProgramContext.declareProof({ Proof, proof: this });
+    const ProofClass = this.constructor as Subclass<typeof ProofBase>;
+    ZkProgramContext.declareProof({ ProofClass, proofInstance: this });
     return true;
   }
 
@@ -115,6 +119,15 @@ class ProofBase<Input = any, Output = any> {
   }
   publicFields() {
     return (this.constructor as typeof ProofBase).publicFields(this);
+  }
+
+  static _proofFromBase64(proofString: string, maxProofsVerified: 0 | 1 | 2) {
+    assertBindingsInitialized();
+    return Pickles.proofOfBase64(proofString, maxProofsVerified)[1];
+  }
+  static _proofToBase64(proof: Pickles.Proof, maxProofsVerified: 0 | 1 | 2) {
+    assertBindingsInitialized();
+    return Pickles.proofToBase64([maxProofsVerified, proof]);
   }
 }
 
@@ -451,4 +464,11 @@ function extractProofTypes(type: ProvableType) {
   let value = ProvableType.synthesize(type);
   let proofValues = extractProofs(value);
   return proofValues.map((proof) => proof.constructor as typeof ProofBase);
+}
+
+function assertBindingsInitialized() {
+  assert(
+    areBindingsInitialized,
+    'Bindings are not initialized. Try calling `await initializeBindings()` first.'
+  );
 }